stockmaster/requirements.txt
# Django & Extensions
Django==4.2.7
django-environ==0.11.2
django-cors-headers==4.3.0
django-filter==23.3
django-storages==1.14.2
django-extensions==3.2.3
django-debug-toolbar==4.2.0
whitenoise==6.6.0
python-dotenv==1.0.0

# REST API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Shopify
ShopifyAPI==12.3.0
PyJWT==2.8.0
cryptography==41.0.5

# Time & Date
pytz==2023.3
python-dateutil==2.8.2

# Utils
Pillow==10.1.0
requests==2.31.0
urllib3==2.0.7
python-magic==0.4.27
apscheduler==3.10.4

# Production
gunicorn==21.2.0
sentry-sdk==1.32.0 